requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
web3==5.31.4
eth-account==0.5.9
py-ecc==4.0.0
click==7.0
pycryptodome==3.9.7
